faster mul

+++ src/geometry/imul.h
@@ -6,7 +6,7 @@ namespace geometry {
 
 void initMultiplication();
 
-uint32_t mul8To16(uint16_t a, uint16_t b);
+uint16_t mul8To16(uint8_t a, uint8_t b);
 uint32_t mul16To32(uint16_t a, uint16_t b);
 uint64_t mul32To64(uint32_t a, uint32_t b);
